local-event-notifier/requirements.txt
anthropic>=0.39.0
markdown>=3.7
python-dotenv>=1.0.0
requests>=2.32.0
